CREATE Pagination
This commit builds the pagination of the items view, breaking it into a
number of pages.

However, once a new item is loaded, the cache becomes stale and as of
this commit, there was no solution to deleting part of the cache. Fetching
afresh the items afresh can solve this but defeats the purpose of having
the cache. Deleting the cache removes everything, including what is in
the cart.